# rpc: improve the realtime notify performance by 30%