I'm running 10 websocket clients(mainly to test #27439) to receive the real-time events, it turns out the `rpc.Notify` eat too much CPU, see the profile file below:

<img width="1446" alt="image" src="https://github.com/ethereum/go-ethereum/assets/3627395/233d1af0-06e0-4d05-b3f6-52dede9aedc2">

And found there are too many CPUs used in JSON.marshaling, and after digging in the source code, seems we can improve the performance by reducing 1 JSON.marshaling. So I add a benchmark, here is the result of the benchmark: 

```bash
go test -bench=.  -v -run BenchmarkNotify -count 10
```

> old 

```txt
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/rpc
cpu: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz
BenchmarkNotify
BenchmarkNotify-16         19371             66857 ns/op
BenchmarkNotify-16         19032             59029 ns/op
BenchmarkNotify-16         19263             67036 ns/op
BenchmarkNotify-16         16597             61500 ns/op
BenchmarkNotify-16         19549             65228 ns/op
BenchmarkNotify-16         16611             72045 ns/op
BenchmarkNotify-16         16962             68970 ns/op
BenchmarkNotify-16         16431             66456 ns/op
BenchmarkNotify-16         23557             49392 ns/op
BenchmarkNotify-16         23340             53930 ns/op
PASS
ok      github.com/ethereum/go-ethereum/rpc     18.224s
```

> new

```txt
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/rpc
cpu: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz
BenchmarkNotify
BenchmarkNotify-16         24076             45048 ns/op
BenchmarkNotify-16         25297             42735 ns/op
BenchmarkNotify-16         23407             45985 ns/op
BenchmarkNotify-16         34939             46708 ns/op
BenchmarkNotify-16         36223             45710 ns/op
BenchmarkNotify-16         24590             46140 ns/op
BenchmarkNotify-16         22858             45640 ns/op
BenchmarkNotify-16         26013             47350 ns/op
BenchmarkNotify-16         31420             44238 ns/op
BenchmarkNotify-16         26156             45340 ns/op
PASS
ok      github.com/ethereum/go-ethereum/rpc     17.150s
```

generate the bench differ:

```bash
benchstat old.txt new.txt
old.txt:5: missing iteration count
new.txt:5: missing iteration count
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/rpc
cpu: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz
          │   old.txt    │               new.txt               │
          │    sec/op    │   sec/op     vs base                │
Notify-16   65.84µ ± 18%   45.68µ ± 3%  -30.63% (p=0.000 n=10)

```