@holiman 2023-10-13
The percentages weren't quite right: you just discarded the output, but in reality it will be json-encoded at least one more time, which adds to the overhead (and makes the impact less). 

Aside from that, I noticed that we could theoretically take this idea even further, like this: 

```
diff --git a/rpc/subscription.go b/rpc/subscription.go
index 947767e963..920ad960ce 100644
--- a/rpc/subscription.go
+++ b/rpc/subscription.go
@@ -177,16 +177,24 @@ func (n *Notifier) activate() error {
 }
 
 func (n *Notifier) send(sub *Subscription, data interface{}) error {
-	params, _ := json.Marshal(struct {
+	thing := struct {
 		ID     string      `json:"subscription"`
 		Result interface{} `json:"result,omitempty"`
-	}{ID: string(sub.ID), Result: data})
+	}{ID: string(sub.ID), Result: data}
+
 	ctx := context.Background()
 
-	msg := &jsonrpcMessage{
+	msg := struct {
+		Version string          `json:"jsonrpc,omitempty"`
+		ID      json.RawMessage `json:"id,omitempty"`
+		Method  string          `json:"method,omitempty"`
+		Params  interface{}     `json:"params,omitempty"`
+		Error   *jsonError      `json:"error,omitempty"`
+		Result  json.RawMessage `json:"result,omitempty"`
+	}{
 		Version: vsn,
 		Method:  n.namespace + notificationMethodSuffix,
-		Params:  params,
+		Params:  thing,
 	}
 	return n.h.conn.writeJSON(ctx, msg, false)
 }
```
That is, always avoid putting `json.RawMessage` into structs, and instead keep things in object-form as long as possible in the encoding pipeline. 

I don't know if this is worth doing, but if it is, we should probably do it generally, not just in this particular instance.