@jsvisa 2023-10-13
@holiman thanks for your tips. After using the new benchmark code, there are also 23% improvement.

```
>$ cat old.txt
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/rpc
cpu: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz
BenchmarkNotify
BenchmarkNotify-16         17034             87875 ns/op
BenchmarkNotify-16         14574             80534 ns/op
BenchmarkNotify-16         14750             95860 ns/op
BenchmarkNotify-16         13078             90346 ns/op
BenchmarkNotify-16         14583             74846 ns/op
BenchmarkNotify-16         12854             87367 ns/op
BenchmarkNotify-16         14354             82897 ns/op
BenchmarkNotify-16         14950             76899 ns/op
BenchmarkNotify-16         13184             89349 ns/op
BenchmarkNotify-16         15166             77653 ns/op
PASS
ok      github.com/ethereum/go-ethereum/rpc     21.838s
>$ cat new.txt
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/rpc
cpu: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz
BenchmarkNotify
BenchmarkNotify-16         17634             59332 ns/op
BenchmarkNotify-16         19249             65428 ns/op
BenchmarkNotify-16         18812             71263 ns/op
BenchmarkNotify-16         17366             66708 ns/op
BenchmarkNotify-16         17824             69048 ns/op
BenchmarkNotify-16         17083             64215 ns/op
BenchmarkNotify-16         19155             61733 ns/op
BenchmarkNotify-16         21247             66564 ns/op
BenchmarkNotify-16         22184             63637 ns/op
BenchmarkNotify-16         20374             56697 ns/op
PASS
ok      github.com/ethereum/go-ethereum/rpc     18.847s

>$ benchstat old.txt new.txt
old.txt:5: missing iteration count
new.txt:5: missing iteration count
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/rpc
cpu: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz
          ?   old.txt    ?               new.txt               ?
          ?    sec/op    ?   sec/op     vs base                ?
Notify-16   85.13µ ± 10%   64.82µ ± 8%  -23.86% (p=0.000 n=10)

```