@jsvisa 2023-10-13
> That is, always avoid putting json.RawMessage into structs, and instead keep things in object-form as long as possible in the encoding pipeline.

Try to adapt more to that style, and see if the perform differs