@holiman 2023-10-13
```
[user@work go-ethereum]$ benchstat after_0.txt after_1.txt 
name      old time/op  new time/op  delta
Notify-8  53.0µs ±17%  40.8µs ±14%  -23.08%  (p=0.000 n=10+10)
[user@work go-ethereum]$ benchstat after_1.txt after_2.txt 
name      old time/op  new time/op  delta
Notify-8  40.8µs ±14%  27.1µs ± 4%  -33.54%  (p=0.000 n=10+8)
```

I'm getting first a 23% improv, then another 33%. Not suprisingly