@jsvisa 2023-10-13
@holiman I think we can declare different JSON-RPC structs for the request and response, using `json.RawMessage` for requests, and `interface` for response, wdyt?