@holiman 2023-10-13
Yes, might be a good solution. But this really is @fjl's domains, so don't take my idea as a statement of "this is how we want to do it".