@jsvisa 2023-10-17
Here is a simple script to benchmark the single vs multi `json.Marshal`:


```go
package main

import (
	"encoding/json"
	"testing"
)

type rawResult struct {
	ID     string          `json:"subscription"`
	Result json.RawMessage `json:"result,omitempty"`
}

type rawMessage struct {
	Version string          `json:"jsonrpc,omitempty"`
	ID      json.RawMessage `json:"id,omitempty"`
	Method  string          `json:"method,omitempty"`
	Params  json.RawMessage `json:"params,omitempty"`
	Result  json.RawMessage `json:"result,omitempty"`
}

type interfaceResult struct {
	ID     string      `json:"subscription"`
	Result interface{} `json:"result,omitempty"`
}

type interfaceMessage struct {
	Version string          `json:"jsonrpc,omitempty"`
	ID      json.RawMessage `json:"id,omitempty"`
	Method  string          `json:"method,omitempty"`
	Params  interface{}     `json:"params,omitempty"`
	Result  interface{}     `json:"result,omitempty"`
}

type Header struct {
	Number int
}

func BenchmarkInterface(b *testing.B) {
	header := Header{Number: 1}
	for i := 0; i < b.N; i++ {
		params := interfaceResult{ID: "id", Result: header}
		msg := interfaceMessage{
			Version: "2.0",
			Method:  "rpc_test",
			Params:  params,
		}
		json.Marshal(msg)
	}
}

func BenchmarkRaw(b *testing.B) {
	header := Header{Number: 1}
	for i := 0; i < b.N; i++ {
		enc, _ := json.Marshal(header)
		params, _ := json.Marshal(&rawResult{ID: "id", Result: enc})
		msg := rawMessage{
			Version: "2.0",
			Method:  "rpc_test",
			Params:  params,
		}
		json.Marshal(msg)
	}
}
```

And here is a result in my PC:

```bash
go test -bench=. -v
goos: linux
goarch: amd64
pkg: github.com/jsvisa/gochain
cpu: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz
BenchmarkInterface
BenchmarkInterface-16             682720              1941 ns/op
BenchmarkRaw
BenchmarkRaw-16                   375933              3832 ns/op
PASS
ok      github.com/jsvisa/gochain       4.670s
```