@s1na 2023-10-17
This seems to resonate with https://github.com/golang/go/issues/33422. Unrelated: but there I saw a proposal for `encoding/json/v2` https://github.com/golang/go/discussions/63397. Looking cool!