@s1na 2023-10-17
I'm curious how the performance will be if the parameter passed to `Notify` is not a pointer, e.g. `h types.Header`. I imagine there will be some extra copying happening. That said our current subscriptions pass mostly pointers, with the exceptions of the pending transaction subscription which passes in a `common.Hash`.